CPU configuration UI: several improvements
- Removed 'More' button from the Processor tab in Edit Template.
This was done to be standardized with the Edit Guest tab that
does not have this button to show/hide the max processor
value.

- Save button state is now being controlled by the init_processor_tab
function (at least inside the Processor tab).

- 'Current CPU' is now editable at all times, in both Edit Guest
and Edit Template dialogs, even when defining a topology. Inserting
a wrong value in the field will invalidate the save state, disabling
the button. When defining a topology and Current CPU has an
inconsistent value, it will be automatically adjusted to be (1)
a multiple of threads or (2) equal to new Max CPU value, in that
order.

- When not defining a topology, setting any Current CPU value greater
than Max CPU will force the field back to the Max CPU value.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>