Edit Guest dialog: fixing 'Save' button on Processor tab
Edit Guest does not work the same way as the Edit
Template dialog - each tab has its own submit function
and it is not possible to submit multiple tabs in the
same dialog.

This patch splits the logic to update the CPU configuration
that was contained in the 'generalSubmit' function and put it
into a new function called 'processorSubmit'. This function
will be fired when the 'Save' button is pressed when the
'Processor' tab is active.

This patch also further enhances the control of the 'Save'
button state in the 'init_processor_tab' JS function by
making it always enabled when there is no topology being set.
If current CPU is set to a value above max CPU it will be
rewrite to be equal to max CPU; if max CPU is set to a value
below current CPU, current CPU is set to max CPU.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>